pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llll"
version = "0.1.0"
description = "LEGO lin la loop — an agentic development loop for LEGO Mindstorms via Pybricks"
requires-python = ">=3.10"
dependencies = [
    "pybricksdev>=2.3.2",
    "mcp>=1.0.0",
]

[project.scripts]
llll = "llll.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/llll"]

[tool.hatch.build.targets.wheel.sources]
src = ""
